+++ test_autofit/interpolator/conftest.py
@@ -1,5 +1,7 @@
 import pytest
+
 import autofit as af
+from autofit.interpolator import CovarianceInterpolator
 
 
 @pytest.fixture(name="model_instance")
@@ -23,3 +25,37 @@ def make_instances(model_instance):
             )
         ),
     ]
+
+
+@pytest.fixture(name="interpolator")
+def make_interpolator():
+    samples_list = [
+        af.SamplesPDF(
+            model=af.Collection(
+                t=value,
+                gaussian=af.Model(
+                    af.Gaussian,
+                    centre=af.GaussianPrior(mean=1.0, sigma=1.0),
+                    normalization=af.GaussianPrior(mean=1.0, sigma=1.0),
+                    sigma=af.GaussianPrior(mean=1.0, sigma=1.0),
+                ),
+            ),
+            sample_list=[
+                af.Sample(
+                    log_likelihood=-i,
+                    log_prior=1.0,
+                    weight=1.0,
+                    kwargs={
+                        ("gaussian", "centre"): value + i,
+                        ("gaussian", "normalization"): value + i**2,
+                        ("gaussian", "sigma"): value + i**3,
+                    },
+                )
+                for i in range(3)
+            ],
+        )
+        for value in range(3)
+    ]
+    return CovarianceInterpolator(
+        samples_list,
+    )
